Adding angio model and layers segmentation

+++ HCT-9000/modules/NormData/NormQuadNFL.cpp
@@ -280,7 +280,7 @@ int NormData::NormQuadNFL::calculateWeightedRegionPercentile(int sector, const s
 		}
 	}
 
-	int totalCount = tsnitIndices.size();
+	int totalCount = (int)tsnitIndices.size();
 	float redRatio = (float)redCount / totalCount;
 	float abnormalRatio = (float)(redCount + yellowCount) / totalCount;
